rpi4: Add el3 cpu frequency trampoline
The rpi4 has a unused(?) "ARM mailbox" interface that can
be used as a generic memory region that generates interrupts
when written. Wire up the first interrupt a group0/secure and
when we detect it pick up the requested frequency and create
a videocore mailbox call to adjust the frequency.

The result of this, is that we can implement ACPI CPPC
frequency scaling (actually it might be possible to do
legacy pstate/etc too) on the rpi4.

Signed-off-by: Jeremy Linton <jeremy.linton@arm.com>
Change-Id: Iecf1dc8351b31e205e760883d56d40087c27a5f1