rpi4: Re-enable PCC channel to provide proper relative perf
CPCC has ref and delivered perf counters that need to be read
in order to compute a clock rate. The RPI doesn't have hardware
to do this. Instead we setup a PCC conduit to el3 where we take
over the DT region and fill it out with the queried registers.

Signed-off-by: Jeremy Linton <jeremy.linton@arm.com>