plat/rpi4: Add POE fan object on a pcc channel
This is a rough bit of code which attaches the POE
hat mailbox interface to a generic PCC channel that
should in theory be callable from AML. Thus we can use
a variable speed fan.

At this moment this is a WIP because it turns out the AML
to PCC definition in the spec isn't clear to me. The
trigger and results of a pcc operation are a little hand
wavy, and while I can probably poke the mailbox interface
directly, I would like to see if there is OS level support
for the pcc channel rather than doing all the handshaking
in AML.

If that turns out to be a longer term project we could just
bypass PCC and trigger the TFA->mailbox calls directly in AML
and leave the current fan speed in a known location.

Signed-off-by: Jeremy Linton <jeremy.linton@arm.com>
Change-Id: I6675e53a6d31f4dd72b3f80d65e070e72d055aab